Change `fitsInFloat`, `fitsInDouble` to follow the changes in `asFloat`, `asDouble`.

+++ truffle/src/com.oracle.truffle.api.interop/src/com/oracle/truffle/api/interop/DefaultIntegerExports.java
@@ -75,7 +75,7 @@ final class DefaultIntegerExports {
     static boolean fitsInFloat(Integer receiver) {
         int i = receiver;
         float f = i;
-        return f == i;
+        return (int) f == i;
     }
 
     @ExportMessage
